+++ src/PeriodFormattingTest.php
@@ -35,10 +35,11 @@ final class PeriodFormattingTest extends TestCase
     {
         date_default_timezone_set('Africa/Nairobi');
         $period = Period::fromDate(new DateTimeImmutable('2014-05-01'), new DateTimeImmutable('2014-05-08'));
-        $res = $period->toIso8601();
+        $format = 'Y-m-d\TH:i:s';
 
-        self::assertTrue(str_contains($res, '2014-04-30T21:00:00'));
-        self::assertTrue(str_contains($res, '2014-05-07T21:00:00'));
+        self::assertSame($period->toIso8601($format), '2014-04-30T21:00:00/2014-05-07T21:00:00');
+        self::assertSame($period->toIso80000($format), '[2014-05-01T00:00:00, 2014-05-08T00:00:00)');
+        self::assertSame($period->toBourbaki($format), '[2014-05-01T00:00:00, 2014-05-08T00:00:00[');
     }
 
     public function testJsonSerialize(): void
@@ -60,10 +61,10 @@ final class PeriodFormattingTest extends TestCase
     public function testFormat(): void
     {
         date_default_timezone_set('Africa/Nairobi');
-        self::assertSame('[2015-04, 2015-05)', Period::fromMonth(2015, 4)->toNotation('Y-m'));
+        self::assertSame('[2015-04, 2015-05)', Period::fromMonth(2015, 4)->toIso80000('Y-m'));
         self::assertSame(
             '[2015-04-01 Africa/Nairobi, 2015-04-01 Africa/Nairobi)',
-            (Period::fromDate(new DateTimeImmutable('2015-04-01'), new DateTimeImmutable('2015-04-01')))->toNotation('Y-m-d e')
+            (Period::fromDate(new DateTimeImmutable('2015-04-01'), new DateTimeImmutable('2015-04-01')))->toIso80000('Y-m-d e')
         );
     }
 }